+++ app/src/main/java/me/blog/korn123/easydiary/domain/model/History.kt
@@ -4,5 +4,5 @@ data class History(
     val historyTag: String,
     val date: String,
     val attachedPhotoPath: String,
-    val sequence: Int
+    val sequence: Long,
 )
